refactor: move random functions back to random_ file
